refactor: Schedule Dto 전용 dto/schedule 패키지 추가
